Feature/37 email domain management

+++ arn-backend/src/main/java/com/ukf/arn/Universities/UniversityRepository.java
@@ -3,6 +3,6 @@ package com.ukf.arn.Universities;
 import com.ukf.arn.Entities.University;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UniversityRepository extends JpaRepository<University, Integer> {
+public interface UniversityRepository extends JpaRepository<University, Long> {
 
 }
